python_src/requirements.txt
fastapi
uvicorn
pydantic
transformers
pymilvus
numpy
python-dotenv
torch
scikit-learn
bson
pymongo
google-generativeai
google-genai